Replace local_ordinal_type with native Fortran int
Also simplify literal types in the test code: this cleans up a number of
warnings about conversions, including an unintentional conversion to
single-precision reals (the literal 3.0 was used).

Also commented out unused functions that were generating warnings.

+++ src/tpetra/tutorials/MapsAndVectors.F90
@@ -29,8 +29,8 @@ type(TpetraMultiVector) :: x, y, z
 
 ! -- Scalars
 integer :: ierr
-integer(size_type) :: my_rank, num_procs, k
-integer(size_type) :: num_local_entries, num_elements_per_proc
+integer :: my_rank, num_procs, k
+integer :: num_local_entries, num_elements_per_proc
 integer(global_size_type) :: num_global_entries
 logical :: zero_out
 real(scalar_type) :: alpha, beta, gamma
